+++ app/lib/ui/components/switch/switch_item_demo_screen.dart
@@ -30,9 +30,9 @@ import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_switch.
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_textfield.dart';
 import 'package:ouds_flutter_demo/ui/utilities/detail_screen_header.dart';
 import 'package:ouds_flutter_demo/ui/utilities/dismiss_keyboard.dart';
+import 'package:ouds_flutter_demo/ui/utilities/light_dark_box.dart';
 import 'package:ouds_flutter_demo/ui/utilities/reference_design_version_component.dart';
 import 'package:ouds_flutter_demo/ui/utilities/sheets_bottom/ouds_sheets_bottom.dart';
-import 'package:ouds_flutter_demo/ui/utilities/theme_colored_box.dart';
 import 'package:ouds_theme_contract/ouds_component_version.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
 import 'package:provider/provider.dart';
@@ -147,65 +147,31 @@ class _SwitchButtonItemDemoState extends State<_SwitchButtonItemDemo> {
     customizationState = ControlItemCustomization.of(context);
     themeController = Provider.of<ThemeController>(context, listen: true);
 
-    return Column(
-      children: [
-        ThemeBox(
-          hasConstrainedMaxWidthOption: true,
-          themeContract: themeController!.currentTheme,
-          themeMode: themeController!.isInverseDarkTheme ? ThemeMode.light : ThemeMode.dark,
-          child: Column(
-            children: [
-              OudsSwitchButtonItem(
-                value: _isSwitchOn,
-                onChanged: customizationState!.hasEnabled
-                    ? (bool? newValue) {
-                  setState(() {
-                    _isSwitchOn = newValue!;
-                  });
-                }
-                    : null,
-                title: ControlItemCustomizationUtils.getLabelText(customizationState!),
-                helperTitle: ControlItemCustomizationUtils.getHelperLabelText(customizationState!),
-                reversed: customizationState!.hasReversed ? true : false,
-                readOnly: customizationState!.hasReadOnly ? true : false,
-                icon: customizationState!.hasIcon ? AppAssets.icons.functionalSocialAndEngagementHeartEmpty(themeController!) : null,
-                isError: customizationState!.hasError ? true : false,
-                errorText: ControlItemCustomizationUtils.getErrorMessageLabelText(customizationState!),
-                divider: customizationState!.hasDivider ? true : false,
-                constrainedMaxWidth: customizationState!.hasConstrainedMaxWidth ? true : false,
-              ),
-            ],
+    return LightDarkBox(
+      hasConstrainedMaxWidthOption: true,
+      child: Column(
+        children: [
+          OudsSwitchButtonItem(
+            value: _isSwitchOn,
+            onChanged: customizationState!.hasEnabled
+                ? (bool? newValue) {
+                    setState(() {
+                      _isSwitchOn = newValue!;
+                    });
+                  }
+                : null,
+            title: ControlItemCustomizationUtils.getLabelText(customizationState!),
+            helperTitle: ControlItemCustomizationUtils.getHelperLabelText(customizationState!),
+            reversed: customizationState!.hasReversed ? true : false,
+            readOnly: customizationState!.hasReadOnly ? true : false,
+            icon: customizationState!.hasIcon ? AppAssets.icons.functionalSocialAndEngagementHeartEmpty(themeController!) : null,
+            isError: customizationState!.hasError ? true : false,
+            errorText: ControlItemCustomizationUtils.getErrorMessageLabelText(customizationState!),
+            divider: customizationState!.hasDivider ? true : false,
+            constrainedMaxWidth: customizationState!.hasConstrainedMaxWidth ? true : false,
           ),
-        ),
-        ThemeBox(
-          hasConstrainedMaxWidthOption: true,
-          themeContract: themeController!.currentTheme,
-          themeMode: themeController!.isInverseDarkTheme ? ThemeMode.dark : ThemeMode.light,
-          child: Column(
-            children: [
-              OudsSwitchButtonItem(
-                value: _isSwitchOn,
-                onChanged: customizationState!.hasEnabled
-                    ? (bool? newValue) {
-                  setState(() {
-                    _isSwitchOn = newValue!;
-                  });
-                }
-                    : null,
-                title: ControlItemCustomizationUtils.getLabelText(customizationState!),
-                helperTitle: ControlItemCustomizationUtils.getHelperLabelText(customizationState!),
-                reversed: customizationState!.hasReversed ? true : false,
-                readOnly: customizationState!.hasReadOnly ? true : false,
-                icon: customizationState!.hasIcon ? AppAssets.icons.functionalSocialAndEngagementHeartEmpty(themeController!) : null,
-                isError: customizationState!.hasError ? true : false,
-                errorText: ControlItemCustomizationUtils.getErrorMessageLabelText(customizationState!),
-                divider: customizationState!.hasDivider ? true : false,
-                constrainedMaxWidth: customizationState!.hasConstrainedMaxWidth ? true : false,
-              ),
-            ],
-          ),
-        ),
-      ],
+        ],
+      ),
     );
   }
 }